Capture profit after time limit expires

+++ cpp/sandbox/src/alpaca/types.cpp
@@ -65,7 +65,7 @@ struct quote_t {
   double bid = 0.00;
   double price = 0.00;
   std::string symbol = "";
-  long int timestamp;
+  int timestamp;
 };
 } // namespace t
 } // namespace Alpaca
